add dialog of modal (delete)

+++ src/app/courses/containers/courses/courses.component.ts
@@ -7,6 +7,7 @@ import { ErrorDialogComponent } from 'src/app/shared/components/error-dialog/err
 
 import { Course } from '../../model/course';
 import { CoursesService } from '../../services/courses.service';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/error-dialog/confirmation-dialog/confirmation-dialog.component';
 
 
 @Component({
@@ -60,13 +61,22 @@ onEdit(course: Course){
 }
 
 onRemove(course: Course){
-  this.coursesService.remove(course._id).subscribe(
-    () =>{
-      this.refresh();
-      this.snackBar.open('Curso removido com sucesso!.', 'X', {duration: 5000, verticalPosition: 'top', horizontalPosition: 'center'});
-    },
-    () => this.onError('Erro ao tentar remover curso')
-  )
-}
 
+  const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+    data: 'Tem certeza que deseja remover este curso ?',
+  });
+
+  dialogRef.afterClosed().subscribe( (result: boolean) => {
+    if(result){
+      this.coursesService.remove(course._id).subscribe(
+        () =>{
+          this.refresh();
+          this.snackBar.open('Curso removido com sucesso!.', 'X', {duration: 5000, verticalPosition: 'top', horizontalPosition: 'center'});
+        },
+        () => this.onError('Erro ao tentar remover curso')
+      );
+    }
+  });
+
+ }
 }
